requirements.txt
Django==5.2.1
djangorestframework==3.14.0
Pillow==10.2.0  # For image handling
python-dotenv==1.0.1  # For environment variables
django-cors-headers==4.3.1  # For handling CORS
django-filter==24.1  # For filtering API results
qrcode==7.4.2  # For QR code generation 